+++ src/hommx/helpers.py
@@ -17,7 +17,6 @@ import pyvista as pv
 import ufl
 from dolfinx import fem, mesh, plot
 from dolfinx.fem.petsc import LinearProblem
-from dolfinx_mpc import MultiPointConstraint
 from mpi4py import MPI
 from petsc4py import PETSc
 from petsc4py.PETSc import ScalarType
@@ -208,293 +207,6 @@ def rescale_mesh_in_place(
     msh.geometry.x[:] *= scale
     msh.geometry.x[:] += shift
     return msh
-
-
-def create_periodic_boundary_conditions(
-    function_space: fem.FunctionSpace,
-    bcs: list[fem.DirichletBC] | None = None,
-) -> MultiPointConstraint:
-    """Creates periodic boundary condition on the unit square or unit cube.
-    For implementation details see
-    [`_create_periodic_boundary_conditions_2d`][hommx.helpers._create_periodic_boundary_conditions_2d]
-    and
-    [`_create_periodic_boundary_conditions_3d`][hommx.helpers._create_periodic_boundary_conditions_3d]
-    """
-    msh = function_space.mesh
-    if msh.topology.dim == 1:
-        raise ValueError("Periodic boundary conditions in 1d not implemented.")
-    if msh.topology.dim == 2:
-        return _create_periodic_boundary_conditions_2d(msh, function_space, bcs)
-    if msh.topology.dim == 3:
-        return _create_periodic_boundary_conditions_3d(msh, function_space, bcs)
-    raise ValueError(
-        f"Unkown topology dimension. {function_space.mesh.topology.dim=} is something unexpected"
-    )
-
-
-def _create_periodic_boundary_conditions_2d(
-    msh: mesh.Mesh,
-    function_space: fem.FunctionSpace,
-    bcs: list[fem.DirichletBC] | None = None,
-) -> MultiPointConstraint:
-    """Creates periodic boundary condition on the unit square using dolfinx_mpc.
-
-    This is done using create_periodic_constraints.
-    We do this by forcing $u(x, y_1) = u(x, y_0)$ and $u(x_1, y) = u(x_0, y)$,
-    on the unit square we would have $(x_0, x_1) = (y_0, y_1) = (0, 1)$
-    Internally we create slave DoFs at $u(x, y_1)$ and $u(x_1, y)$.
-    Some attention need to be paid because the node at $(x_1,y_1)$ is constrained twice.
-    We use a workaround adapted from: [dolfinx_mpc](https://github.com/jorgensd/dolfinx_mpc/blob/main/python/demos/demo_periodic_gep.py)
-
-    Args:
-        msh: Mesh that the boundary conditions should be applied to
-        function_space: Function Space on whicht the bc should be applied
-        bcs: Dirichlet Boundary Conditions, here the constraint is ignored
-    """
-    # meshtags used to locate boundary entitites
-    if bcs is None:
-        bcs = []
-    fdim = msh.topology.dim - 1
-    TAG_X = 2
-    TAG_Y = 3
-
-    # figure out master and slave coordinates, i.e. bottom, top, left, right of box
-    slave_coord_x = np.max(msh.geometry.x, axis=0)[0]
-    slave_coord_y = np.max(msh.geometry.x, axis=0)[1]
-
-    master_coord_x = np.min(msh.geometry.x, axis=0)[0]
-    master_coord_y = np.min(msh.geometry.x, axis=0)[1]
-
-    # marker functions for slaves and master
-    def is_slave_x(x):
-        return np.isclose(x[0], slave_coord_x)
-
-    def is_slave_y(x):
-        return np.isclose(x[1], slave_coord_y)
-
-    # relations between master and slave, i.e. u(x, 1) = u(x, 0)
-    def slave_to_master_map_x(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        idx = is_slave_y(x)
-        out_x[0][idx] = np.nan  # don't map double constrained points
-        return out_x
-
-    def slave_to_master_map_y(x):
-        out_x = x.copy()
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        idx = is_slave_x(x)
-        out_x[1][idx] = np.nan  # don't map double constrained points
-        return out_x
-
-    # locate boundary nodes and tag them
-    meshtags = []
-    facets_x = mesh.locate_entities_boundary(msh, fdim, is_slave_x)
-    facets_y = mesh.locate_entities_boundary(msh, fdim, is_slave_y)
-    meshtags = (
-        mesh.meshtags(
-            msh,
-            fdim,
-            facets_x[np.argsort(facets_x)],
-            np.full(len(facets_x), TAG_X, dtype=np.int32),
-        ),
-        mesh.meshtags(
-            msh,
-            fdim,
-            facets_y[np.argsort(facets_y)],
-            np.full(len(facets_y), TAG_Y, dtype=np.int32),
-        ),
-    )
-
-    # create multi-point-constraint
-    mpc = MultiPointConstraint(function_space)
-    # map boundaries
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[0], TAG_X, slave_to_master_map_x, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[1], TAG_Y, slave_to_master_map_y, bcs
-    )
-
-    # map (1, 1) to (0, 0)
-    def slave_to_master_map(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        idx = np.logical_and(is_slave_x(x), is_slave_y(x))
-        out_x[0][~idx] = np.nan
-        out_x[1][~idx] = np.nan
-        return out_x
-
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[1], TAG_Y, slave_to_master_map, bcs
-    )
-    mpc.finalize()
-    return mpc
-
-
-def _create_periodic_boundary_conditions_3d(
-    msh: mesh.Mesh,
-    function_space: fem.FunctionSpace,
-    bcs: list[fem.DirichletBC] | None = None,
-) -> MultiPointConstraint:
-    """Creates periodic boundary condition on the unit cube using dolfinx_mpc.
-
-    This is done using create_periodic_constraints.
-    We do this by forcing $u(x, y_1, z) = u(x, y_0, z)$ and so on for all directions
-    on the unit square we would have $(z_0, z_1) = (x_0, x_1) = (y_0, y_1) = (0, 1)$
-    Internally we create slave DoFs at $u(x, y_1, z)$ and $u(x_1, y, z)$ etc.
-    Some attention need to be paid because the nodes at $(x_1, y_1, z)$ are constrained twice etc for other edges
-    And $(x_1, y_1, z_1)$ is constrained three times.
-    We use a workaround adapted from: [dolfinx_mpc](https://github.com/jorgensd/dolfinx_mpc/blob/main/python/demos/demo_periodic_gep.py)
-
-    Args:
-        msh: Mesh that the boundary conditions should be applied to
-        function_space: Function Space on which the bc should be applied
-        bcs: Dirichlet Boundary Conditions, here the constraint is ignored
-    """
-    # meshtags used to locate boundary entitites
-    if bcs is None:
-        bcs = []
-    fdim = msh.topology.dim - 1
-    TAG_X = 2
-    TAG_Y = 3
-    TAG_Z = 4
-
-    # figure out master and slave coordinates, i.e. bottom, top, left, right of box
-    slave_coord_x = np.max(msh.geometry.x, axis=0)[0]
-    slave_coord_y = np.max(msh.geometry.x, axis=0)[1]
-    slave_coord_z = np.max(msh.geometry.x, axis=0)[2]
-
-    master_coord_x = np.min(msh.geometry.x, axis=0)[0]
-    master_coord_y = np.min(msh.geometry.x, axis=0)[1]
-    master_coord_z = np.min(msh.geometry.x, axis=0)[2]
-
-    # marker functions for slaves and master
-    def is_slave_x(x):
-        return np.isclose(x[0], slave_coord_x)
-
-    def is_slave_y(x):
-        return np.isclose(x[1], slave_coord_y)
-
-    def is_slave_z(x):
-        return np.isclose(x[2], slave_coord_z)
-
-    # locate boundary nodes and tag them
-    meshtags = []
-    facets_x = mesh.locate_entities_boundary(msh, fdim, is_slave_x)
-    facets_y = mesh.locate_entities_boundary(msh, fdim, is_slave_y)
-    facets_z = mesh.locate_entities_boundary(msh, fdim, is_slave_z)
-    meshtags = (
-        mesh.meshtags(
-            msh,
-            fdim,
-            facets_x[np.argsort(facets_x)],
-            np.full(len(facets_x), TAG_X, dtype=np.int32),
-        ),
-        mesh.meshtags(
-            msh,
-            fdim,
-            facets_y[np.argsort(facets_y)],
-            np.full(len(facets_y), TAG_Y, dtype=np.int32),
-        ),
-        mesh.meshtags(
-            msh,
-            fdim,
-            facets_z[np.argsort(facets_z)],
-            np.full(len(facets_z), TAG_Z, dtype=np.int32),
-        ),
-    )
-
-    # relations between master and slave that are constrained once, i.e. u(x, 1, z) = u(x, 0, z)
-    def slave_to_master_map_x(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        idx = is_slave_y(x) | is_slave_z(x)
-        out_x[0][idx] = np.nan  # don't map double constrained points
-        return out_x
-
-    def slave_to_master_map_y(x):
-        out_x = x.copy()
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        idx = is_slave_x(x) | is_slave_z(x)
-        out_x[1][idx] = np.nan  # don't map double constrained points
-        return out_x
-
-    def slave_to_master_map_z(x):
-        out_x = x.copy()
-        out_x[2] = x[2] - (slave_coord_z - master_coord_z)
-        idx = is_slave_x(x) | is_slave_y(x)
-        out_x[2][idx] = np.nan  # don't map double constrained points
-        return out_x
-
-    # create multi-point-constraint
-    mpc = MultiPointConstraint(function_space)
-    # map boundaries
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[0], TAG_X, slave_to_master_map_x, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[1], TAG_Y, slave_to_master_map_y, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[2], TAG_Z, slave_to_master_map_z, bcs
-    )
-
-    # map (1, 1, z) to (0, 0, z), (x, 1, 1) to (x, 0, 0) etc., i.e. double constrained
-    def slave_to_master_map_x_y(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        idx = is_slave_x(x) & is_slave_y(x) & ~is_slave_z(x)
-        out_x[0][~idx] = np.nan
-        out_x[1][~idx] = np.nan
-        return out_x
-
-    def slave_to_master_map_x_z(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        out_x[2] = x[2] - (slave_coord_z - master_coord_z)
-        idx = is_slave_x(x) & is_slave_z(x) & ~is_slave_y(x)
-        out_x[0][~idx] = np.nan
-        out_x[2][~idx] = np.nan
-        return out_x
-
-    def slave_to_master_map_y_z(x):
-        out_x = x.copy()
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        out_x[2] = x[2] - (slave_coord_z - master_coord_z)
-        idx = is_slave_y(x) & is_slave_z(x) & ~is_slave_x(x)
-        out_x[1][~idx] = np.nan
-        out_x[2][~idx] = np.nan
-        return out_x
-
-    # map (1,1,1) to (0,0,0), i.e. three times constrained
-    def slave_to_master_map_x_y_z(x):
-        out_x = x.copy()
-        out_x[0] = x[0] - (slave_coord_x - master_coord_x)
-        out_x[1] = x[1] - (slave_coord_y - master_coord_y)
-        out_x[2] = x[2] - (slave_coord_z - master_coord_z)
-        idx = is_slave_y(x) & is_slave_z(x) & is_slave_x(x)
-        out_x[0][~idx] = np.nan
-        out_x[1][~idx] = np.nan
-        out_x[2][~idx] = np.nan
-        return out_x
-
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[0], TAG_X, slave_to_master_map_x_y, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[0], TAG_X, slave_to_master_map_x_z, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[1], TAG_Y, slave_to_master_map_y_z, bcs
-    )
-    mpc.create_periodic_constraint_topological(
-        function_space, meshtags[1], TAG_Y, slave_to_master_map_x_y_z, bcs
-    )
-    mpc.finalize()
-    return mpc
 
 
 class PoissonFEM:
