+++ src/hommx/hmm.py
@@ -15,7 +15,6 @@ from petsc4py import PETSc
 from tqdm import tqdm
 
 import hommx.cell_problem as cell_problem
-import hommx.helpers as helpers
 
 
 def _triangle_area(points):
@@ -180,7 +179,7 @@ class BaseHMM(ABC):
         """Setup cell problem specifics"""
         # micro function space and periodic boundary conditions
         self._V_micro = self._setup_micro_function_space()
-        self._mpc = helpers.create_periodic_boundary_conditions(self._V_micro, self._bcs)
+        self._mpc = cell_problem.create_periodic_boundary_conditions(self._V_micro, self._bcs)
         self._points_micro = self._V_micro.tabulate_dof_coordinates()
         self._v_tilde = ufl.TrialFunction(self._V_micro)
         self._z = ufl.TestFunction(self._V_micro)
@@ -1096,7 +1095,7 @@ class BasePeriodicHMM(ABC):
         self._u = fem.Function(self._V_macro)
 
         self._V_micro = self._setup_micro_function_space()
-        self._mpc = helpers.create_periodic_boundary_conditions(self._V_micro, [])
+        self._mpc = cell_problem.create_periodic_boundary_conditions(self._V_micro, [])
         self._v = ufl.TrialFunction(self._V_micro)
         self._z = ufl.TestFunction(self._V_micro)
         self._y = ufl.SpatialCoordinate(self._cell_mesh)
